Use native storage to avoid localStorage limits on iOS
iOS limits localStorage to 5MB. The frame data files have started to
use just over this, meaning they are no longer saved. This commit
switches to using Capacitor's Storage plugin instead. Because this is
asynchronous, this commit also adds a FrameDataGate component. This
prevents the app from loading until the frame data has been retrieved
from storage, avoiding the need to refactor the whole app to deal with
this change.